+++ NEWS.md
@@ -1,7 +1,8 @@
 # mvnchol 0.2.0
 
 * Add GEFS Innsbruck data.
-* `means()` function to extract means for basic Cholesky.
+* `fam$means()` function to extract means for basic Cholesky.
+* `fam$stdev()` function to extract standard deviations for basic Cholesky.
 
 # mvnchol 0.1.3